Update scheduling to wait for task completion on shutdown
