+++ analysis-platform/src/main/java/com/kenesys/analysisplatform/config/SchedulingConfig.java
@@ -1,6 +1,10 @@
 package com.kenesys.analysisplatform.config;
 
 
+import com.kenesys.analysisplatform.services.templates.GitScanner;
+import com.kenesys.analysisplatform.services.templates.TemplateService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.TaskScheduler;
@@ -13,16 +17,33 @@ import org.springframework.scheduling.config.ScheduledTaskRegistrar;
 @EnableScheduling
 public class SchedulingConfig  implements SchedulingConfigurer {
 
+    @Value("${templates.gitscanner.schedule}")
+    private String gitScannerSchedule;
+
+    @Value( "${templates.gitscanner.gitdir}" )
+    private String gitDirectory;
+
+    @Value( "${templates.gitscanner.gituri}" )
+    private String gitRepositoryUri;
+
+    @Value( "${templates.gitscanner.gitbranch}" )
+    public String gitBranch;
+
+    @Autowired
+    private TemplateService templateService;
+
     @Override
     public void configureTasks(ScheduledTaskRegistrar taskRegistrar) {
-        taskRegistrar.setTaskScheduler(poolScheduler());
+        taskRegistrar.setTaskScheduler(platformScheduler());
     }
 
     @Bean
-    public TaskScheduler poolScheduler() {
+    public TaskScheduler platformScheduler() {
         ThreadPoolTaskScheduler scheduler = new ThreadPoolTaskScheduler();
-        scheduler.setThreadNamePrefix("poolScheduler");
+        scheduler.setThreadNamePrefix("platformScheduler");
         scheduler.setWaitForTasksToCompleteOnShutdown(true);
+        scheduler.initialize();
+        scheduler.scheduleWithFixedDelay(new GitScanner(gitDirectory, gitRepositoryUri, gitBranch, templateService), Long.valueOf(gitScannerSchedule));
         return scheduler;
     }
 }
